+++ spec/support/output/multiple_offenses.rb
@@ -1,0 +1,5 @@
+for number in [*1..100] do
+  number
+end
+
+shuffled = [1, 2, 3].shuffle.first